refactor: move postgre pool logic to `utils::postgre`
